+++ crates/teloxide-core/src/types/mask_position.rs
@@ -1,3 +1,5 @@
+use std::hash::Hash;
+
 use serde::{Deserialize, Serialize};
 
 /// This object describes the position on faces where a mask should be placed by
@@ -5,7 +7,8 @@ use serde::{Deserialize, Serialize};
 ///
 /// [The official docs](https://core.telegram.org/bots/api#maskposition).
 #[serde_with::skip_serializing_none]
-#[derive(Copy, Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Clone, Copy, Debug)]
+#[derive(Serialize, Deserialize)]
 pub struct MaskPosition {
     /// The part of the face relative to which the mask should be placed. One
     /// of `forehead`, `eyes`, `mouth`, or `chin`.
@@ -25,8 +28,38 @@ pub struct MaskPosition {
     pub scale: f64,
 }
 
+impl PartialEq for MaskPosition {
+    fn eq(&self, other: &Self) -> bool {
+        self.point == other.point
+            && self.x_shift.total_cmp(&other.x_shift).is_eq()
+            && self.y_shift.total_cmp(&other.y_shift).is_eq()
+            && self.scale.total_cmp(&other.scale).is_eq()
+    }
+}
+
+impl Eq for MaskPosition {}
+
+impl Hash for MaskPosition {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        self.point.hash(state);
+
+        // Skip values which are NaN and canonicalize negative zero (-0.0 + 0.0 == +0.0)
+        if !self.x_shift.is_nan() {
+            (self.x_shift + 0.0).to_bits().hash(state);
+        }
+        if !self.y_shift.is_nan() {
+            (self.y_shift + 0.0).to_bits().hash(state);
+        }
+        if !self.scale.is_nan() {
+            (self.scale + 0.0).to_bits().hash(state);
+        }
+    }
+}
+
 /// The part of the face relative to which the mask should be placed.
-#[derive(Copy, Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Clone, Copy, Debug)]
+#[derive(PartialEq, Eq, Hash)]
+#[derive(Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum MaskPoint {
     Forehead,